chrome://device-log
New Chromebook which I could not connect to my cable modem to use Ethernet. To connect to the internet, I had to toggle the autoassign IP address of the Ethernet connection profile. THen, the IP I was assign was not one like 192.168.0.2 or 102.169.100.11, but rather an actual foreign address like 174.68.33.XXX